Add files via upload
The Hilbert Transform of the Gaussian focuses on the reactions, transformations, and calculations involved in plasma dispersion function. This notebook first offers information on the properties of Z, including symmetry properties, values for special arguments, power series, asymptotic expansion, and differential equation characterization. The python project notebook then ponders on the applications to plasma physics. Numerical calculations on the function of Z are presented. The python project notebook takes a look at table generation and accuracy wherein various methods are proposed in computing the error function in the multiple regions of the complex plane.